Changed all floats to doubles and changed the relative coordinate direction to degrees_d and distance to millimetres_u

+++ tests/camera_fakes.h
@@ -70,7 +70,7 @@ extern "C" {
 #define CAMERA_FAKES(FAKE) \
     FAKE(gu_camera_equals)
 
-DECLARE_FAKE_VALUE_FUNC(bool, gu_camera_equals, const gu_camera, const gu_camera, const float)
+DECLARE_FAKE_VALUE_FUNC(bool, gu_camera_equals, const gu_camera, const gu_camera, const double)
 
 #ifdef __cplusplus
 }
